+++ ChatHistoryIterator.h
@@ -56,7 +56,7 @@ class ChatHistoryIterator : public Iterator<string*> {
         /**
          * @brief returns the current index
          */
-        size_t getIndex() const override;
+        int getIndex() const override;
 };
 
 #endif // CHATHISTORYITERATOR_H
